Add mkfs_options_extra attribute to Partition
This is laying out support for explicit mkfs extra options in
the various config files repart.d supports.

This commit adds .mkfs_options_extra attribute to the Partition struct.

If not NULL, .mkfs_options_extra is exclusively used to set formatting
options.
If NULL, the current environment-variable based approach is used
instead.

+++ src/partition/repart.c
@@ -295,8 +295,8 @@ typedef struct Partition {
         uint64_t copy_blocks_size;
 
         char *format;
+        char **mkfs_options_extra;
         char **copy_files;
-        char **mkfs_options;
         char **exclude_files_source;
         char **exclude_files_target;
         char **make_directories;
@@ -460,6 +460,7 @@ static Partition* partition_free(Partition *p) {
         safe_close(p->copy_blocks_fd);
 
         free(p->format);
+        strv_free(p->mkfs_options_extra);
         strv_free(p->copy_files);
         strv_free(p->exclude_files_source);
         strv_free(p->exclude_files_target);
@@ -495,6 +496,7 @@ static void partition_foreignize(Partition *p) {
         p->copy_blocks_root = NULL;
 
         p->format = mfree(p->format);
+        p->mkfs_options_extra = strv_free(p->mkfs_options_extra);
         p->copy_files = strv_free(p->copy_files);
         p->exclude_files_source = strv_free(p->exclude_files_source);
         p->exclude_files_target = strv_free(p->exclude_files_target);
@@ -1669,33 +1671,6 @@ static int config_parse_make_dirs(
         }
 }
 
-static int config_parse_mkfs_options(
-                const char *unit,
-                const char *filename,
-                unsigned line,
-                const char *section,
-                unsigned section_line,
-                const char *lvalue,
-                int ltype,
-                const char *rvalue,
-                void *data,
-                void *userdata) {
-                
-        _cleanup_strv_free_ char **l = NULL;
-        char ***opt = ASSERT_PTR(data);
-        const char *p = ASSERT_PTR(rvalue);
-        
-        if (p) {
-                l = strv_split(p, NULL);
-                if (!l)
-                        return log_oom();
-        }
-
-        *opt = TAKE_PTR(l);
-
-        return 0;
-}
-
 static DEFINE_CONFIG_PARSE_ENUM_WITH_DEFAULT(config_parse_encrypt, encrypt_mode, EncryptMode, ENCRYPT_OFF, "Invalid encryption mode");
 
 static int config_parse_gpt_flags(
@@ -1902,7 +1877,6 @@ static int partition_read_definition(Partition *p, const char *path, const char 
                 { "Partition", "FactoryReset",             config_parse_bool,             0, &p->factory_reset           },
                 { "Partition", "CopyBlocks",               config_parse_copy_blocks,      0, p                           },
                 { "Partition", "Format",                   config_parse_fstype,           0, &p->format                  },
-                { "Partition", "MkfsOptions",              config_parse_mkfs_options,     0, &p->mkfs_options            },
                 { "Partition", "CopyFiles",                config_parse_copy_files,       0, &p->copy_files              },
                 { "Partition", "ExcludeFiles",             config_parse_exclude_files,    0, &p->exclude_files_source    },
                 { "Partition", "ExcludeFilesTarget",       config_parse_exclude_files,    0, &p->exclude_files_target    },
@@ -5062,17 +5036,18 @@ static int context_mkfs(Context *context) {
                                 return r;
                 }
 
-                r = mkfs_options_from_env("REPART", p->format, &extra_mkfs_options);
+                r = strv_copy_unless_empty(p->mkfs_options_extra, &extra_mkfs_options);
                 if (r < 0)
-                        return log_error_errno(r,
-                                               "Failed to determine mkfs command line options for '%s': %m",
-                                               p->format);
+                        return log_oom();
                 
-                if (p->mkfs_options) {
-                        r = strv_extend_strv(&extra_mkfs_options, p->mkfs_options, /* filter_duplicates = */ 0);
+                if (!extra_mkfs_options) {
+                        /* Only read environment if mkfs options are not provided in config files */
+                        r = mkfs_options_from_env("REPART", p->format, &extra_mkfs_options);
                         if (r < 0)
-                                return log_error_errno(r, "Failed to append mkfs options.");
-                }
+                                return log_error_errno(r,
+                                                       "Failed to determine mkfs command line options for '%s': %m",
+                                                       p->format);
+                }                                                
                 
                 r = make_filesystem(partition_target_path(t), p->format, strempty(p->new_label), root,
                                     p->fs_uuid, arg_discard, /* quiet = */ false,
@@ -6604,16 +6579,17 @@ static int context_minimize(Context *context) {
                                 return r;
                 }
 
-                r = mkfs_options_from_env("REPART", p->format, &extra_mkfs_options);
+                r = strv_copy_unless_empty(p->mkfs_options_extra, &extra_mkfs_options);
                 if (r < 0)
-                        return log_error_errno(r,
-                                               "Failed to determine mkfs command line options for '%s': %m",
-                                               p->format);
+                        return log_oom();
                 
-                if (p->mkfs_options) {
-                        r = strv_extend_strv(&extra_mkfs_options, p->mkfs_options, /* filter_duplicates = */ 0);
+                if (!extra_mkfs_options) {
+                        /* Only read environment if mkfs options are not provided in config files */
+                        r = mkfs_options_from_env("REPART", p->format, &extra_mkfs_options);
                         if (r < 0)
-                                return log_error_errno(r, "Failed to append mkfs options.");
+                                return log_error_errno(r,
+                                                       "Failed to determine mkfs command line options for '%s': %m",
+                                                       p->format);
                 }
 
                 r = make_filesystem(d ? d->node : temp,
